Add SweetAlert2 library and add window confirmation for delete product

+++ frontend/frontend/src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/common/product';
 import { ProductService } from 'src/app/services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -24,9 +25,30 @@ export class ProductListComponent implements OnInit {
   }
 
   deleteProductById(id: number){
-    this.productService.deleteProductById(id).subscribe(
-      ()=> this.listProducts()
-    );
+
+    Swal.fire({
+      title: "¿Está seguro que desea eliminar el registro?",
+      text: "",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productService.deleteProductById(id).subscribe(
+          ()=> this.listProducts()
+        );
+        Swal.fire({
+          title: "¡Productos!",
+          text: "Producto eliminado correctamente.",
+          icon: "success"
+        });
+      }
+    });
+
+    
   }
